Removed first call of word variable
The first call of the word variable in the first few lines was not necessary. In the first few lines of code this variable randomizes a word from the list once and uses that word for the duration of the program. By moving the variable to the while loop, the variable changes every time the user decides to play a new game.